test functional with logger
